+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-06-18T09:16:55.261Z")
+    created: Optional[datetime] = parse_datetime("2025-06-18T12:05:08.542Z")
     r"""The date and time the resource was created"""