+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JY157KFCH1X3M10DXG0MTN3G"
+        "01JY1EVM45TY01QF13Z9D8ATZV"
     )
     r"""The id of the resource"""